V0.1 - Working basic ticketing system. Further development needed.

+++ app/models/customer.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, Boolean
+from decimal import Decimal
+
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -24,7 +26,7 @@ class Customer(Base):
         ForeignKey("invoice_frequencies.id")
     )
     payment_terms: Mapped[str | None] = mapped_column(String(100))
-    credit_limit: Mapped[float | None] = mapped_column(Numeric(12, 2))
+    credit_limit: Mapped[Decimal | None] = mapped_column(Numeric(12, 2))
     on_stop: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     cash_account: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     do_not_invoice: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
